+++ CHANGES.rst
@@ -6,6 +6,7 @@ Change log for gocept.bbissues
 ================
 
 - Add time_rendered as a variable passed to the jinja2 template.
+- Add the count of comments to the issue/pullrequest and pass to template.
 
 
 0.6 (2016-01-27)